MPG_App_files/requirements.txt
numpy==1.17.2
pandas==0.25.1
streamlit==0.67.1
Pillow==7.2.0
scikit_learn==0.23.2
